chore: add plugin interface alias
