--- plugin/client/interface.go
@@ -1,571 +1,0 @@
-/*
-Copyright 2021 The Katanomi Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package client
-
-import (
-	"context"
-	"io"
-
-	cloudevent "github.com/cloudevents/sdk-go/v2"
-	"github.com/emicklei/go-restful/v3"
-	coderepositoryv1alpha1 "github.com/katanomi/pkg/apis/coderepository/v1alpha1"
-	metav1alpha1 "github.com/katanomi/pkg/apis/meta/v1alpha1"
-	"github.com/katanomi/pkg/plugin/client/base"
-	"go.uber.org/zap"
-	corev1 "k8s.io/api/core/v1"
-	"knative.dev/pkg/apis"
-)
-
-// Interface base interface for plugins
-type Interface interface {
-	Path() string
-	Setup(context.Context, *zap.SugaredLogger) error
-}
-
-// PluginRegister plugin registration methods to update IntegrationClass status
-type PluginRegister interface {
-	Interface
-	PluginAddressable
-	// GetIntegrationClassName returns integration class name
-	GetIntegrationClassName() string
-	// GetWebhookURL Returns a Webhook accessible URL for external tools
-	// If not supported return nil, false
-	GetWebhookURL() (*apis.URL, bool)
-	// GetSupportedVersions Returns a list of supported versions by the plugin
-	// For SaaS platform plugins use a "online" version.
-	GetSupportedVersions() []string
-	// GetSecretTypes Returns all secret types supported by the plugin
-	GetSecretTypes() []string
-	// GetReplicationPolicyTypes return replication policy types for ClusterIntegration
-	GetReplicationPolicyTypes() []string
-	// GetResourceTypes Returns a list of Resource types that can be used in ClusterIntegration and Integration
-	GetResourceTypes() []string
-	// GetAllowEmptySecret Returns if an empty secret is allowed with IntegrationClass
-	GetAllowEmptySecret() []string
-}
-
-// PluginAddressable provides methods to get plugin address url
-type PluginAddressable interface {
-	// GetAddressURL Returns its own plugin access URL
-	GetAddressURL() *apis.URL
-}
-
-// DependentResourceGetter checks and returns dependent resource references
-type DependentResourceGetter interface {
-	// GetDependentResources parses params and returns expected dependent resources
-	GetDependentResources(ctx context.Context, params []metav1alpha1.Param) ([]corev1.ObjectReference, error)
-}
-
-type AdditionalWebhookRegister interface {
-	// GetWebhookSupport get webhook support map
-	GetWebhookSupport() map[metav1alpha1.WebhookEventSupportType][]string
-}
-
-// ResourcePathFormatter implements a formatter for resource path base on different scene
-type ResourcePathFormatter interface {
-	// GetResourcePathFmt resource path format
-	GetResourcePathFmt() map[metav1alpha1.ResourcePathScene]string
-	// GetSubResourcePathFmt resource path format
-	GetSubResourcePathFmt() map[metav1alpha1.ResourcePathScene]string
-}
-
-// AuthChecker implements an authorization check method for plugins
-type AuthChecker interface {
-	AuthCheck(ctx context.Context, option metav1alpha1.AuthCheckOptions) (*metav1alpha1.AuthCheck, error)
-}
-
-// AuthTokenGenerator implements token generation/refresh API method
-type AuthTokenGenerator interface {
-	AuthToken(ctx context.Context) (*metav1alpha1.AuthToken, error)
-}
-
-// ProjectLister list project api
-type ProjectLister interface {
-	Interface
-	ListProjects(ctx context.Context, option metav1alpha1.ListOptions) (*metav1alpha1.ProjectList, error)
-}
-
-type PluginAttributes interface {
-	SetAttribute(k string, values ...string)
-	GetAttribute(k string) []string
-	Attributes() map[string][]string
-}
-
-// PluginVersionAttributes get diff configurations for different versions.
-type PluginVersionAttributes interface {
-	// GetVersionAttributes get the differential configuration of the specified version
-	GetVersionAttributes(version string) map[string][]string
-
-	// SetVersionAttributes set the differential configuration of the specified version
-	SetVersionAttributes(version string, attributes map[string][]string)
-}
-
-// ProjectGetter list project api
-type ProjectGetter interface {
-	Interface
-	GetProject(ctx context.Context, id string) (*metav1alpha1.Project, error)
-}
-
-// ProjectCreator create project api
-type ProjectCreator interface {
-	Interface
-	CreateProject(ctx context.Context, project *metav1alpha1.Project) (*metav1alpha1.Project, error)
-}
-
-// ProjectDeleter create project api
-type ProjectDeleter interface {
-	Interface
-	DeleteProject(ctx context.Context, project *metav1alpha1.Project) error
-}
-
-// RepositoryLister list repository
-type RepositoryLister interface {
-	Interface
-	ListRepositories(
-		ctx context.Context,
-		params metav1alpha1.RepositoryOptions,
-		option metav1alpha1.ListOptions,
-	) (*metav1alpha1.RepositoryList, error)
-}
-
-// RepositoryGetter get repository
-type RepositoryGetter interface {
-	Interface
-	GetRepository(
-		ctx context.Context,
-		params metav1alpha1.RepositoryOptions,
-	) (*metav1alpha1.Repository, error)
-}
-
-// ArtifactLister list artifact
-type ArtifactLister interface {
-	Interface
-	ListArtifacts(ctx context.Context, params metav1alpha1.ArtifactOptions, option metav1alpha1.ListOptions) (*metav1alpha1.ArtifactList, error)
-}
-
-// ProjectArtifactLister list project-level artifacts
-type ProjectArtifactLister interface {
-	Interface
-	ListProjectArtifacts(ctx context.Context, params metav1alpha1.ArtifactOptions, option metav1alpha1.ListOptions) (*metav1alpha1.ArtifactList, error)
-}
-
-// ArtifactGetter get artifact detail
-type ArtifactGetter interface {
-	Interface
-	GetArtifact(ctx context.Context, params metav1alpha1.ArtifactOptions) (*metav1alpha1.Artifact, error)
-}
-
-// ProjectArtifactGetter get artifact detail
-type ProjectArtifactGetter interface {
-	Interface
-	GetProjectArtifact(ctx context.Context, params metav1alpha1.ProjectArtifactOptions) (*metav1alpha1.Artifact, error)
-}
-
-// ProjectArtifactFileGetter download artifact within a project
-type ProjectArtifactFileGetter interface {
-	Interface
-	GetProjectArtifactFile(ctx context.Context, params metav1alpha1.ProjectArtifactOptions) (io.ReadCloser, error)
-}
-
-// ProjectArtifactDeleter delete artifact
-type ProjectArtifactDeleter interface {
-	Interface
-	DeleteProjectArtifact(ctx context.Context, params metav1alpha1.ProjectArtifactOptions) error
-}
-
-// ProjectArtifactUploader upload artifact
-type ProjectArtifactUploader interface {
-	Interface
-	UploadArtifact(ctx context.Context, params metav1alpha1.ProjectArtifactOptions, r io.Reader) error
-}
-
-// ArtifactDeleter delete artifact
-type ArtifactDeleter interface {
-	Interface
-	DeleteArtifact(ctx context.Context, params metav1alpha1.ArtifactOptions) error
-}
-
-// ArtifactTagDeleter delete a specific tag of the artifact.
-type ArtifactTagDeleter interface {
-	Interface
-	DeleteArtifactTag(ctx context.Context, params metav1alpha1.ArtifactTagOptions) error
-}
-
-// ScanImage scan image
-type ScanImage interface {
-	Interface
-	ScanImage(ctx context.Context, params metav1alpha1.ArtifactOptions) error
-}
-
-// ImageConfigGetter get image config
-type ImageConfigGetter interface {
-	Interface
-	GetImageConfig(ctx context.Context, params metav1alpha1.ArtifactOptions) (*metav1alpha1.ImageConfig, error)
-}
-
-// WebhookRegister used to register and manage webhooks
-type WebhookRegister interface {
-	// Use the methods below to manage webhooks in the target platform
-	CreateWebhook(ctx context.Context, spec metav1alpha1.WebhookRegisterSpec, secret corev1.Secret) (metav1alpha1.WebhookRegisterStatus, error)
-	UpdateWebhook(ctx context.Context, spec metav1alpha1.WebhookRegisterSpec, secret corev1.Secret) (metav1alpha1.WebhookRegisterStatus, error)
-	DeleteWebhook(ctx context.Context, spec metav1alpha1.WebhookRegisterSpec, secret corev1.Secret) error
-	ListWebhooks(ctx context.Context, uri apis.URL, secret corev1.Secret) ([]metav1alpha1.WebhookRegisterStatus, error)
-}
-
-// GitTriggerRegister used to register GitTrigger
-// TODO: need refactor: maybe integration plugin should decided how to generate cloudevents filters
-// up to now, it is not a better solution that relying on plugins to give some events type to GitTriggerReconcile.
-//
-// PullRequestCloudEventFilter() CloudEventFilters
-// BranchCloudEventFilter() CloudEventFilters
-// TagCloudEventFilter() CloudEventFilters
-// WebHook() WebHook
-type GitTriggerRegister interface {
-	GetIntegrationClassName() string
-
-	// cloud event type of pull request hook that will match
-	PullRequestEventType() string
-
-	// cloud event type of push hook that will match
-	PushEventType() string
-
-	// cloud event type of push hook that will match
-	TagEventType() string
-}
-
-// WebhookResourceDiffer used to compare different webhook resources in order to provide
-// a way to merge webhook registration requests. If not provided, the resource's URI will be directly compared
-type WebhookResourceDiffer interface {
-	// IsSameResource will provide two ResourceURI
-	// the plugin should discern if they are the same.
-	// If this method is not implemented a standard comparisons will be used
-	IsSameResource(ctx context.Context, i, j metav1alpha1.ResourceURI) bool
-}
-
-// WebhookReceiver receives a webhook request with validation and transform it into a cloud event
-type WebhookReceiver interface {
-	Interface
-	ReceiveWebhook(ctx context.Context, req *restful.Request, secret string) (cloudevent.Event, error)
-}
-
-// GitPullRequestCommentCreator create pull request comment functions
-type GitPullRequestCommentCreator interface {
-	Interface
-	CreatePullRequestComment(ctx context.Context, option metav1alpha1.CreatePullRequestCommentPayload) (metav1alpha1.GitPullRequestNote, error)
-}
-
-// GitPullRequestCommentUpdater updates pull request comment
-type GitPullRequestCommentUpdater interface {
-	Interface
-	UpdatePullRequestComment(ctx context.Context, option metav1alpha1.UpdatePullRequestCommentPayload) (metav1alpha1.GitPullRequestNote, error)
-}
-
-// GitPullRequestCommentLister list pull request comment functions
-type GitPullRequestCommentLister interface {
-	Interface
-	ListPullRequestComment(
-		ctx context.Context,
-		option metav1alpha1.GitPullRequestOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitPullRequestNoteList, error)
-}
-
-// GitPullRequestHandler list, get and create pr function
-type GitPullRequestHandler interface {
-	Interface
-	ListGitPullRequest(
-		ctx context.Context,
-		option metav1alpha1.GitPullRequestListOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitPullRequestList, error)
-	GetGitPullRequest(ctx context.Context, option metav1alpha1.GitPullRequestOption) (metav1alpha1.GitPullRequest, error)
-	CreatePullRequest(ctx context.Context, payload metav1alpha1.CreatePullRequestPayload) (metav1alpha1.GitPullRequest, error)
-}
-
-// GitCommitGetter get git commit
-type GitCommitGetter interface {
-	Interface
-	GetGitCommit(ctx context.Context, option metav1alpha1.GitCommitOption) (metav1alpha1.GitCommit, error)
-}
-
-// GitCommitCreator create git commit
-type GitCommitCreator interface {
-	Interface
-	CreateGitCommit(ctx context.Context, option coderepositoryv1alpha1.CreateGitCommitOption) (metav1alpha1.GitCommit, error)
-}
-
-// GitCommitLister List git commit
-type GitCommitLister interface {
-	Interface
-	ListGitCommit(
-		ctx context.Context,
-		option metav1alpha1.GitCommitListOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitCommitList, error)
-}
-
-// GitBranchLister List git branch
-type GitBranchLister interface {
-	Interface
-	ListGitBranch(ctx context.Context, branchOption metav1alpha1.GitBranchOption, option metav1alpha1.ListOptions) (metav1alpha1.GitBranchList, error)
-}
-
-// GitBranchGetter get git branch
-type GitBranchGetter interface {
-	Interface
-	GetGitBranch(ctx context.Context, repoOption metav1alpha1.GitRepo, branch string) (metav1alpha1.GitBranch, error)
-}
-
-// GitBranchCreator create git branch,github, gogs don't support create branch
-type GitBranchCreator interface {
-	Interface
-	CreateGitBranch(ctx context.Context, payload metav1alpha1.CreateBranchPayload) (metav1alpha1.GitBranch, error)
-}
-
-// GitRepoFileGetter used to get a file content
-type GitRepoFileGetter interface {
-	Interface
-	GetGitRepoFile(ctx context.Context, option metav1alpha1.GitRepoFileOption) (metav1alpha1.GitRepoFile, error)
-}
-
-// GitRepoFileCreator used to create a file, gogs don't support
-type GitRepoFileCreator interface {
-	Interface
-	CreateGitRepoFile(ctx context.Context, payload metav1alpha1.CreateRepoFilePayload) (metav1alpha1.GitCommit, error)
-}
-
-// GitRepositoryCreator create a git repository
-type GitRepositoryCreator interface {
-	Interface
-	CreateGitRepository(ctx context.Context, payload metav1alpha1.CreateGitRepositoryPayload) (metav1alpha1.GitRepository, error)
-}
-
-// GitRepositoryDeleter delete a git repository
-type GitRepositoryDeleter interface {
-	Interface
-	DeleteGitRepository(ctx context.Context, gitRepo metav1alpha1.GitRepo) error
-}
-
-// GitRepositoryLister list git repository
-type GitRepositoryLister interface {
-	Interface
-	ListGitRepository(
-		ctx context.Context,
-		id, keyword string,
-		subtype metav1alpha1.ProjectSubType,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitRepositoryList, error)
-}
-
-// GitRepositoryGetter get git repository
-type GitRepositoryGetter interface {
-	Interface
-	GetGitRepository(ctx context.Context, repoOption metav1alpha1.GitRepo) (metav1alpha1.GitRepository, error)
-}
-
-// GitRepositoryFileTreeGetter get git repository file tree
-type GitRepositoryFileTreeGetter interface {
-	Interface
-	GetGitRepositoryFileTree(
-		ctx context.Context,
-		repoOption metav1alpha1.GitRepoFileTreeOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitRepositoryFileTree, error)
-}
-
-// GitCommitStatusLister list git commit status
-type GitCommitStatusLister interface {
-	Interface
-	ListGitCommitStatus(
-		ctx context.Context,
-		option metav1alpha1.GitCommitOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitCommitStatusList, error)
-}
-
-// GitCommitStatusCreator create git commit status
-type GitCommitStatusCreator interface {
-	Interface
-	CreateGitCommitStatus(ctx context.Context, payload metav1alpha1.CreateCommitStatusPayload) (metav1alpha1.GitCommitStatus, error)
-}
-
-// GitCommitCommentLister list git commit comment
-type GitCommitCommentLister interface {
-	Interface
-	ListGitCommitComment(
-		ctx context.Context,
-		option metav1alpha1.GitCommitOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitCommitCommentList, error)
-}
-
-// GitCommitCommentCreator create git commit comment
-type GitCommitCommentCreator interface {
-	Interface
-	CreateGitCommitComment(ctx context.Context, payload metav1alpha1.CreateCommitCommentPayload) (metav1alpha1.GitCommitComment, error)
-}
-
-// GitRepositoryTagCreator create git repository tag
-type GitRepositoryTagCreator interface {
-	Interface
-	CreateGitRepositoryTag(ctx context.Context, option metav1alpha1.CreateGitTagPayload) (metav1alpha1.GitRepositoryTag, error)
-}
-
-// GitRepositoryTagGetter get git repository Tag
-type GitRepositoryTagGetter interface {
-	Interface
-	GetGitRepositoryTag(
-		ctx context.Context,
-		option metav1alpha1.GitRepositoryTagOption,
-	) (metav1alpha1.GitRepositoryTag, error)
-}
-
-// GitRepositoryTagLister list git repository Tag
-type GitRepositoryTagLister interface {
-	Interface
-	ListGitRepositoryTag(
-		ctx context.Context,
-		option metav1alpha1.GitRepositoryTagListOption,
-		listOption metav1alpha1.ListOptions,
-	) (metav1alpha1.GitRepositoryTagList, error)
-}
-
-type CodeQualityGetter interface {
-	Interface
-	GetCodeQuality(ctx context.Context, projectKey string) (*metav1alpha1.CodeQuality, error)
-	GetCodeQualityOverviewByBranch(ctx context.Context, opt metav1alpha1.CodeQualityBaseOption) (*metav1alpha1.CodeQuality, error)
-	GetCodeQualityLineCharts(ctx context.Context, opt metav1alpha1.CodeQualityLineChartOption) (*metav1alpha1.CodeQualityLineChart, error)
-	GetOverview(ctx context.Context) (*metav1alpha1.CodeQualityProjectOverview, error)
-	GetSummaryByTaskID(ctx context.Context, opt metav1alpha1.CodeQualityTaskOption) (*metav1alpha1.CodeQualityTaskMetrics, error)
-}
-
-type BlobStoreLister interface {
-	Interface
-	ListBlobStores(ctx context.Context, listOption metav1alpha1.ListOptions) (*metav1alpha1.BlobStoreList, error)
-}
-
-// Client inteface for PluginClient, client code shoud use the interface
-// as dependency
-type Client = base.Client
-
-type ClientProjectGetter interface {
-	Project(meta Meta, secret corev1.Secret) ClientProject
-}
-
-// ArtifactTriggerRegister used to register ArtifactTrigger
-type ArtifactTriggerRegister interface {
-	GetIntegrationClassName() string
-
-	// cloud event type of push hook that will match
-	PushEventType() string
-}
-
-// IssueLister issue lister
-type IssueLister interface {
-	Interface
-	ListIssues(ctx context.Context, params metav1alpha1.IssueOptions, option metav1alpha1.ListOptions) (*metav1alpha1.IssueList, error)
-}
-
-type IssueGetter interface {
-	Interface
-	GetIssue(ctx context.Context, params metav1alpha1.IssueOptions, option metav1alpha1.ListOptions) (*metav1alpha1.Issue, error)
-}
-
-type IssueBranchLister interface {
-	Interface
-	ListIssueBranches(ctx context.Context, params metav1alpha1.IssueOptions, option metav1alpha1.ListOptions) (*metav1alpha1.BranchList, error)
-}
-
-type IssueBranchCreator interface {
-	Interface
-	CreateIssueBranch(ctx context.Context, params metav1alpha1.IssueOptions, payload metav1alpha1.Branch) (*metav1alpha1.Branch, error)
-}
-
-type IssueBranchDeleter interface {
-	Interface
-	DeleteIssueBranch(ctx context.Context, params metav1alpha1.IssueOptions, option metav1alpha1.ListOptions) error
-}
-
-type IssueAttributeGetter interface {
-	Interface
-	GetIssueAttribute(ctx context.Context, params metav1alpha1.IssueOptions, option metav1alpha1.ListOptions) (*metav1alpha1.Attribute, error)
-}
-
-type ProjectUserLister interface {
-	Interface
-	ListProjectUsers(ctx context.Context, params metav1alpha1.UserOptions, option metav1alpha1.ListOptions) (*metav1alpha1.UserList, error)
-}
-
-// TestPlanLister list test plans
-type TestPlanLister interface {
-	Interface
-	ListTestPlans(ctx context.Context, params metav1alpha1.TestProjectOptions, option metav1alpha1.ListOptions) (*metav1alpha1.TestPlanList, error)
-}
-
-// TestPlanGetter get a test plan
-type TestPlanGetter interface {
-	Interface
-	GetTestPlan(ctx context.Context, params metav1alpha1.TestProjectOptions) (*metav1alpha1.TestPlan, error)
-}
-
-// TestCaseLister list test cases
-type TestCaseLister interface {
-	Interface
-	ListTestCases(ctx context.Context, params metav1alpha1.TestProjectOptions, options metav1alpha1.ListOptions) (*metav1alpha1.TestCaseList, error)
-}
-
-// TestCaseGetter get a test case
-type TestCaseGetter interface {
-	Interface
-	GetTestCase(ctx context.Context, params metav1alpha1.TestProjectOptions) (*metav1alpha1.TestCase, error)
-}
-
-// TestModuleLister list a test module
-type TestModuleLister interface {
-	Interface
-	ListTestModules(ctx context.Context, params metav1alpha1.TestProjectOptions, options metav1alpha1.ListOptions) (*metav1alpha1.TestModuleList, error)
-}
-
-// TestCaseExecutionLister list test case executions
-type TestCaseExecutionLister interface {
-	Interface
-	ListTestCaseExecutions(ctx context.Context, params metav1alpha1.TestProjectOptions, options metav1alpha1.ListOptions) (*metav1alpha1.TestCaseExecutionList, error)
-}
-
-// TestCaseExecutionCreator create a new test case execution
-type TestCaseExecutionCreator interface {
-	Interface
-	CreateTestCaseExecution(ctx context.Context, params metav1alpha1.TestProjectOptions, payload metav1alpha1.TestCaseExecution) (*metav1alpha1.TestCaseExecution, error)
-}
-
-// LivenessChecker check the tool service is alive
-type LivenessChecker interface {
-	// CheckAlive check the tool service is alive
-	CheckAlive(ctx context.Context) error
-}
-
-// Initializer initialize the tool service
-type Initializer interface {
-	// Initialize  the tool service if desired
-	Initialize(ctx context.Context) error
-}
-
-// ToolMetadataGetter get the version information corresponding to the address.
-type ToolMetadataGetter interface {
-	// GetToolMetadata get the version information corresponding to the address.
-	GetToolMetadata(ctx context.Context) (*metav1alpha1.ToolMeta, error)
-}
