+++ plugin/client/project.go
@@ -20,8 +20,13 @@ import (
 	"context"
 
 	metav1alpha1 "github.com/katanomi/pkg/apis/meta/v1alpha1"
+	corev1 "k8s.io/api/core/v1"
 	duckv1 "knative.dev/pkg/apis/duck/v1"
 )
+
+type ClientProjectGetter interface {
+	Project(meta Meta, secret corev1.Secret) ClientProject
+}
 
 type ClientProject interface {
 	List(ctx context.Context, baseURL *duckv1.Addressable, options ...OptionFunc) (*metav1alpha1.ProjectList, error)
