+++ plugin/types/project.go
@@ -1,0 +1,47 @@
+/*
+Copyright 2024 The Katanomi Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package types
+
+import (
+	"context"
+
+	metav1alpha1 "github.com/katanomi/pkg/apis/meta/v1alpha1"
+)
+
+// ProjectLister list project api
+type ProjectLister interface {
+	Interface
+	ListProjects(ctx context.Context, option metav1alpha1.ListOptions) (*metav1alpha1.ProjectList, error)
+}
+
+// ProjectGetter list project api
+type ProjectGetter interface {
+	Interface
+	GetProject(ctx context.Context, id string) (*metav1alpha1.Project, error)
+}
+
+// ProjectCreator create project api
+type ProjectCreator interface {
+	Interface
+	CreateProject(ctx context.Context, project *metav1alpha1.Project) (*metav1alpha1.Project, error)
+}
+
+// ProjectDeleter create project api
+type ProjectDeleter interface {
+	Interface
+	DeleteProject(ctx context.Context, project *metav1alpha1.Project) error
+}
