+++ plugin/types/repository.go
@@ -1,0 +1,42 @@
+/*
+Copyright 2024 The Katanomi Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package types
+
+import (
+	"context"
+
+	metav1alpha1 "github.com/katanomi/pkg/apis/meta/v1alpha1"
+)
+
+// RepositoryLister list repository
+type RepositoryLister interface {
+	Interface
+	ListRepositories(
+		ctx context.Context,
+		params metav1alpha1.RepositoryOptions,
+		option metav1alpha1.ListOptions,
+	) (*metav1alpha1.RepositoryList, error)
+}
+
+// RepositoryGetter get repository
+type RepositoryGetter interface {
+	Interface
+	GetRepository(
+		ctx context.Context,
+		params metav1alpha1.RepositoryOptions,
+	) (*metav1alpha1.Repository, error)
+}
